串口命令优化，起码能设置正负速度了，电机功能优化

+++ Hardware/Serial.h
@@ -1,8 +1,10 @@
 #ifndef __SERIAL_H
 #define __SERIAL_H
 
+#include "stm32f10x.h"
 
 void Serial_Init(void);
-void USART_Send_Data(int16_t speed1,int16_t speed2,int32_t pos1,int32_t pos2);
+
+void USART_Send_Data(int16_t speed1, int16_t speed2, long pos1, long pos2);
 
 #endif
